+++ CHANGELOG.md
@@ -5,7 +5,7 @@ https://keepachangelog.com/en/1.0.0/
 - `PointlessCollectionToString`: Supports immutable collections
 - `StructWithoutElementaryMethodsOverridden`: No longer generates the elementary methods in both declarations if it is a `partial struct`
 - `StructWithoutElementaryMethodsOverridden`: Rewritten to use `ISymbol`
-- `PointlessCollectionToString`, `ThrowNull`, `SwitchDoesNotHandleAllEnumOptions`, `TestMethodWithoutPublicModifier`, `StructWithoutElementaryMethodsOverridden` and `UnboundedStackalloc` tests use `VerifyCS`
+- `PointlessCollectionToString`, `ThrowNull`, `ExplicitEnumValues`, `SwitchDoesNotHandleAllEnumOptions`, `TestMethodWithoutPublicModifier`, `StructWithoutElementaryMethodsOverridden` and `UnboundedStackalloc` tests use `VerifyCS`
 
 ## [1.19.0] - 2023-01-12
 - `FormReadSynchronously`: Synchronously accessed `HttpRequest.Form` which uses sync-over-async. Use `HttpRequest.ReadFormAsync()` instead