+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 https://keepachangelog.com/en/1.0.0/
 
-## [1.19.1] - 2023-01-12
+## [1.19.1] - 2023-01-14
 - `PointlessCollectionToString`: Supports immutable collections
 - `StructWithoutElementaryMethodsOverridden`: No longer generates the elementary methods in both declarations if it is a `partial struct`
 - `StructWithoutElementaryMethodsOverridden`: Rewritten to use `ISymbol`